Task 1: repo cleanup + shared AppNav layout
- Delete unused HelloWorld.vue scaffold
- Extract duplicated header/nav into shared AppNav.vue (Dashboard,
  Add Transaction, Reports, Logout) with active-link primary style
- Render AppNav from App.vue on authenticated routes only; keep
  simple guest header for login/register
- Add CLAUDE.md project context and PHASE-1-SPEC.md
